fix: 누락된 롤링페이퍼 생성 버튼 추가
