+++ frontend/src/pages/TeamDetailPage/index.tsx
@@ -1,3 +1,4 @@
+import { useNavigate } from "react-router-dom";
 import styled from "@emotion/styled";
 
 import useValidateParam from "@/hooks/useValidateParam";
@@ -10,6 +11,7 @@ import TeamJoinSection from "@/pages/TeamDetailPage/components/TeamJoinSection";
 
 const TeamDetailPage = () => {
   const teamId = useValidateParam<number>("teamId");
+  const navigate = useNavigate();
 
   const { isLoading: isLoadingTeamDetail, data: teamDetail } =
     useReadTeamDetail(teamId);
@@ -31,6 +33,13 @@ const TeamDetailPage = () => {
         color={teamDetail.color}
         joined={teamDetail.joined}
       />
+      <StyledRollingpaperCreateButton
+        onClick={() => {
+          navigate(`/rollingpaper/new?team-id=${teamId}`);
+        }}
+      >
+        📜 롤링페이퍼 만들기
+      </StyledRollingpaperCreateButton>
       {teamDetail.joined ? (
         <RollingpaperList />
       ) : (
@@ -48,6 +57,19 @@ const StyledMain = styled.main`
   gap: 40px;
 
   padding: 28px 0;
+`;
+
+const StyledRollingpaperCreateButton = styled.button`
+  width: 90%;
+  align-self: center;
+  padding: 16px;
+
+  font-size: 18px;
+  background: antiquewhite;
+
+  &:hover {
+    background: #f9e1c3;
+  }
 `;
 
 export default TeamDetailPage;
